+++ common/src/main/java/com/st0x0ef/stellaris/common/utils/Utils.java
@@ -439,6 +439,5 @@ public class Utils {
 
         if (attributeInstance != null)
             attributeInstance.setBaseValue(value);
-
     }
 }