+++ changelog.md
@@ -1,4 +1,4 @@
 # 1.4.14 changelog
 
 ## Fixes
--  Fix armors and shaders
+-  Fix armors rendering with shaders